test(evm): split evmante handlers 1 handler 1 file
