+++ app/evmante_verify_eth_acc_test.go
@@ -189,6 +189,3 @@ func (s *TestSuite) TestAnteDecEthGasConsume() {
 		})
 	}
 }
-
-// func (s *TestSuite) TestAnteDecoratorVerifyEthAcc_CheckTx() {
-// }
